Move CommittOffsets to Committer as it is part of its interface
