+++ zio-kafka/src/main/scala/zio/kafka/consumer/internal/LiveCommitter.scala
@@ -1,15 +1,15 @@
 package zio.kafka.consumer.internal
-import org.apache.kafka.clients.consumer.{ OffsetAndMetadata, OffsetCommitCallback }
+import org.apache.kafka.clients.consumer.{OffsetAndMetadata, OffsetCommitCallback}
 import org.apache.kafka.common.TopicPartition
 import org.apache.kafka.common.errors.RebalanceInProgressException
 import zio.kafka.consumer.Consumer.CommitTimeout
-import zio.kafka.consumer.diagnostics.{ DiagnosticEvent, Diagnostics }
-import zio.kafka.consumer.internal.LiveCommitter.{ Commit, CommitOffsets }
-import zio.{ durationLong, Chunk, Duration, Exit, Promise, Queue, Ref, Runtime, Scope, Task, UIO, Unsafe, ZIO }
+import zio.kafka.consumer.diagnostics.{DiagnosticEvent, Diagnostics}
+import zio.kafka.consumer.internal.Committer.CommitOffsets
+import zio.kafka.consumer.internal.LiveCommitter.Commit
+import zio.{Chunk, Duration, Exit, Promise, Queue, Ref, Runtime, Scope, Task, UIO, Unsafe, ZIO, durationLong}
 
-import java.lang.Math.max
 import java.util
-import java.util.{ Map => JavaMap }
+import java.util.{Map => JavaMap}
 import scala.collection.mutable
 import scala.jdk.CollectionConverters._
 
@@ -164,47 +164,6 @@ private[internal] object LiveCommitter {
   ) {
     @inline def isDone: UIO[Boolean]    = cont.isDone
     @inline def isPending: UIO[Boolean] = isDone.negate
-  }
-
-  // package private for unit testing
-  private[internal] final case class CommitOffsets(offsets: Map[TopicPartition, Long]) {
-
-    /** Returns an estimate of the total offset increase, and a new `CommitOffsets` with the given offsets added. */
-    def addCommits(c: Chunk[Commit]): (Long, CommitOffsets) = {
-      val updatedOffsets = mutable.Map.empty[TopicPartition, Long]
-      updatedOffsets.sizeHint(offsets.size)
-      updatedOffsets ++= offsets
-      var offsetIncrease = 0L
-      c.foreach { commit =>
-        commit.offsets.foreach { case (tp, offsetAndMeta) =>
-          val offset = offsetAndMeta.offset()
-          val maxOffset = updatedOffsets.get(tp) match {
-            case Some(existingOffset) =>
-              offsetIncrease += max(0L, offset - existingOffset)
-              max(existingOffset, offset)
-            case None =>
-              // This partition was not committed to from this consumer yet. Therefore we do not know the offset
-              // increase. A good estimate would be the poll size for this consumer, another okayish estimate is 0.
-              // Lets go with the simplest for now: ```offsetIncrease += 0```
-              offset
-          }
-          updatedOffsets += tp -> maxOffset
-        }
-      }
-      (offsetIncrease, CommitOffsets(offsets = updatedOffsets.toMap))
-    }
-
-    def keepPartitions(tps: Set[TopicPartition]): CommitOffsets =
-      CommitOffsets(offsets.filter { case (tp, _) => tps.contains(tp) })
-
-    def contains(tp: TopicPartition, offset: Long): Boolean =
-      offsets.get(tp).exists(_ >= offset)
-
-    def get(tp: TopicPartition): Option[Long] = offsets.get(tp)
-  }
-
-  private[internal] object CommitOffsets {
-    val empty: CommitOffsets = CommitOffsets(Map.empty)
   }
 
 }
